% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coagulation.R
\docType{data}
\name{coagulation}
\alias{coagulation}
\title{data set}
\format{\preformatted{ chr "coagulation"
}}
\usage{
coagulation
}
\description{
return the coagulation data set
}
\details{
An image file from the coagulation data set of faraway library
}
\keyword{datasets}

