% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oneway.R
\name{oneway.default}
\alias{oneway.default}
\alias{oneway.factor}
\title{oneway}
\usage{
\method{oneway}{default}(z, ...)

\method{oneway}{factor}(z, y, ...)
}
\arguments{
\item{...}{other arguments}

\item{Z}{a list of (possibly) named components}
}
\description{
Estimae the relevant parameters
}
\details{
Estimae the relevant parameters in a one-way AOV with possibly unequal sample sizes.
}

